refactor: split landing page in more components
